Refactor tests to share ordering test data
