+++ src/Shared/Microsoft.AspNetCore.Razor.Utilities.Shared.Test/ReadOnlyListExtensionsTests.cs
@@ -278,332 +278,4 @@ public class ReadOnlyListExtensionsTest
         public static CustomReadOnlyList Create(ReadOnlySpan<int> span)
             => new(span);
     }
-
-    private static Comparison<int> OddBeforeEven
-        => (x, y) => (x % 2 != 0, y % 2 != 0) switch
-        {
-            (true, false) => -1,
-            (false, true) => 1,
-            _ => x.CompareTo(y)
-        };
-
-    public readonly record struct ValueHolder(int Value)
-    {
-        public static implicit operator ValueHolder(int value)
-            => new(value);
-    }
-
-    public static TheoryData<IReadOnlyList<int>, ImmutableArray<int>> OrderTestData
-        => new()
-        {
-            { [1, 2, 3, 4, 5, 6, 7, 8, 9, 10], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10] },
-            { [10, 9, 8, 7, 6, 5, 4, 3, 2, 1], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10] },
-            { [1, 3, 5, 7, 9, 2, 4, 6, 8, 10], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10] },
-            { [2, 5, 8, 1, 3, 9, 7, 4, 10, 6], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10] },
-            { [6, 10, 4, 7, 9, 3, 1, 8, 5, 2], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10] },
-        };
-
-    public static TheoryData<IReadOnlyList<int>, ImmutableArray<int>> OrderTestData_OddBeforeEven
-        => new()
-        {
-            { [1, 2, 3, 4, 5, 6, 7, 8, 9, 10], [1, 3, 5, 7, 9, 2, 4, 6, 8, 10] },
-            { [10, 9, 8, 7, 6, 5, 4, 3, 2, 1], [1, 3, 5, 7, 9, 2, 4, 6, 8, 10] },
-            { [1, 3, 5, 7, 9, 2, 4, 6, 8, 10], [1, 3, 5, 7, 9, 2, 4, 6, 8, 10] },
-            { [2, 5, 8, 1, 3, 9, 7, 4, 10, 6], [1, 3, 5, 7, 9, 2, 4, 6, 8, 10] },
-            { [6, 10, 4, 7, 9, 3, 1, 8, 5, 2], [1, 3, 5, 7, 9, 2, 4, 6, 8, 10] },
-        };
-
-    public static TheoryData<IReadOnlyList<int>, ImmutableArray<int>> OrderDescendingTestData
-        => new()
-        {
-            { [1, 2, 3, 4, 5, 6, 7, 8, 9, 10], [10, 9, 8, 7, 6, 5, 4, 3, 2, 1] },
-            { [10, 9, 8, 7, 6, 5, 4, 3, 2, 1], [10, 9, 8, 7, 6, 5, 4, 3, 2, 1] },
-            { [1, 3, 5, 7, 9, 2, 4, 6, 8, 10], [10, 9, 8, 7, 6, 5, 4, 3, 2, 1] },
-            { [2, 5, 8, 1, 3, 9, 7, 4, 10, 6], [10, 9, 8, 7, 6, 5, 4, 3, 2, 1] },
-            { [6, 10, 4, 7, 9, 3, 1, 8, 5, 2], [10, 9, 8, 7, 6, 5, 4, 3, 2, 1] },
-        };
-
-    public static TheoryData<IReadOnlyList<int>, ImmutableArray<int>> OrderDescendingTestData_OddBeforeEven
-        => new()
-        {
-            { [1, 2, 3, 4, 5, 6, 7, 8, 9, 10], [10, 8, 6, 4, 2, 9, 7, 5, 3, 1] },
-            { [10, 9, 8, 7, 6, 5, 4, 3, 2, 1], [10, 8, 6, 4, 2, 9, 7, 5, 3, 1] },
-            { [1, 3, 5, 7, 9, 2, 4, 6, 8, 10], [10, 8, 6, 4, 2, 9, 7, 5, 3, 1] },
-            { [2, 5, 8, 1, 3, 9, 7, 4, 10, 6], [10, 8, 6, 4, 2, 9, 7, 5, 3, 1] },
-            { [6, 10, 4, 7, 9, 3, 1, 8, 5, 2], [10, 8, 6, 4, 2, 9, 7, 5, 3, 1] },
-        };
-
-    public static TheoryData<IReadOnlyList<ValueHolder>, ImmutableArray<ValueHolder>> OrderByTestData
-        => new()
-        {
-            { [1, 2, 3, 4, 5, 6, 7, 8, 9, 10], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10] },
-            { [10, 9, 8, 7, 6, 5, 4, 3, 2, 1], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10] },
-            { [1, 3, 5, 7, 9, 2, 4, 6, 8, 10], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10] },
-            { [2, 5, 8, 1, 3, 9, 7, 4, 10, 6], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10] },
-            { [6, 10, 4, 7, 9, 3, 1, 8, 5, 2], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10] },
-        };
-
-    public static TheoryData<IReadOnlyList<ValueHolder>, ImmutableArray<ValueHolder>> OrderByTestData_OddBeforeEven
-        => new()
-        {
-            { [1, 2, 3, 4, 5, 6, 7, 8, 9, 10], [1, 3, 5, 7, 9, 2, 4, 6, 8, 10] },
-            { [10, 9, 8, 7, 6, 5, 4, 3, 2, 1], [1, 3, 5, 7, 9, 2, 4, 6, 8, 10] },
-            { [1, 3, 5, 7, 9, 2, 4, 6, 8, 10], [1, 3, 5, 7, 9, 2, 4, 6, 8, 10] },
-            { [2, 5, 8, 1, 3, 9, 7, 4, 10, 6], [1, 3, 5, 7, 9, 2, 4, 6, 8, 10] },
-            { [6, 10, 4, 7, 9, 3, 1, 8, 5, 2], [1, 3, 5, 7, 9, 2, 4, 6, 8, 10] },
-        };
-
-    public static TheoryData<IReadOnlyList<ValueHolder>, ImmutableArray<ValueHolder>> OrderByDescendingTestData
-        => new()
-        {
-            { [1, 2, 3, 4, 5, 6, 7, 8, 9, 10], [10, 9, 8, 7, 6, 5, 4, 3, 2, 1] },
-            { [10, 9, 8, 7, 6, 5, 4, 3, 2, 1], [10, 9, 8, 7, 6, 5, 4, 3, 2, 1] },
-            { [1, 3, 5, 7, 9, 2, 4, 6, 8, 10], [10, 9, 8, 7, 6, 5, 4, 3, 2, 1] },
-            { [2, 5, 8, 1, 3, 9, 7, 4, 10, 6], [10, 9, 8, 7, 6, 5, 4, 3, 2, 1] },
-            { [6, 10, 4, 7, 9, 3, 1, 8, 5, 2], [10, 9, 8, 7, 6, 5, 4, 3, 2, 1] },
-        };
-
-    public static TheoryData<IReadOnlyList<ValueHolder>, ImmutableArray<ValueHolder>> OrderByDescendingTestData_OddBeforeEven
-        => new()
-        {
-            { [1, 2, 3, 4, 5, 6, 7, 8, 9, 10], [10, 8, 6, 4, 2, 9, 7, 5, 3, 1] },
-            { [10, 9, 8, 7, 6, 5, 4, 3, 2, 1], [10, 8, 6, 4, 2, 9, 7, 5, 3, 1] },
-            { [1, 3, 5, 7, 9, 2, 4, 6, 8, 10], [10, 8, 6, 4, 2, 9, 7, 5, 3, 1] },
-            { [2, 5, 8, 1, 3, 9, 7, 4, 10, 6], [10, 8, 6, 4, 2, 9, 7, 5, 3, 1] },
-            { [6, 10, 4, 7, 9, 3, 1, 8, 5, 2], [10, 8, 6, 4, 2, 9, 7, 5, 3, 1] },
-        };
-
-    [Theory]
-    [MemberData(nameof(OrderTestData))]
-    public void OrderAsArray(IReadOnlyList<int> data, ImmutableArray<int> expected)
-    {
-        var sorted = data.OrderAsArray();
-        Assert.Equal<int>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderTestData_OddBeforeEven))]
-    public void OrderAsArray_OddBeforeEven(IReadOnlyList<int> data, ImmutableArray<int> expected)
-    {
-        var sorted = data.OrderAsArray(OddBeforeEven);
-        Assert.Equal<int>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderDescendingTestData))]
-    public void OrderDescendingAsArray(IReadOnlyList<int> data, ImmutableArray<int> expected)
-    {
-        var sorted = data.OrderDescendingAsArray();
-        Assert.Equal<int>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderDescendingTestData_OddBeforeEven))]
-    public void OrderDescendingAsArray_OddBeforeEven(IReadOnlyList<int> data, ImmutableArray<int> expected)
-    {
-        var sorted = data.OrderDescendingAsArray(OddBeforeEven);
-        Assert.Equal<int>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderByTestData))]
-    public void OrderByAsArray(IReadOnlyList<ValueHolder> data, ImmutableArray<ValueHolder> expected)
-    {
-        var sorted = data.OrderByAsArray(static x => x.Value);
-        Assert.Equal<ValueHolder>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderByTestData_OddBeforeEven))]
-    public void OrderByAsArray_OddBeforeEven(IReadOnlyList<ValueHolder> data, ImmutableArray<ValueHolder> expected)
-    {
-        var sorted = data.OrderByAsArray(static x => x.Value, OddBeforeEven);
-        Assert.Equal<ValueHolder>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderByDescendingTestData))]
-    public void OrderByDescendingAsArray(IReadOnlyList<ValueHolder> data, ImmutableArray<ValueHolder> expected)
-    {
-        var sorted = data.OrderByDescendingAsArray(static x => x.Value);
-        Assert.Equal<ValueHolder>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderByDescendingTestData_OddBeforeEven))]
-    public void OrderByDescendingAsArray_OddBeforeEven(IReadOnlyList<ValueHolder> data, ImmutableArray<ValueHolder> expected)
-    {
-        var sorted = data.OrderByDescendingAsArray(static x => x.Value, OddBeforeEven);
-        Assert.Equal<ValueHolder>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderTestData))]
-    public void OrderAsArray_Enumerable(IReadOnlyList<int> data, ImmutableArray<int> expected)
-    {
-        var enumerable = (IEnumerable<int>)data;
-        var sorted = enumerable.OrderAsArray();
-        Assert.Equal<int>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderTestData_OddBeforeEven))]
-    public void OrderAsArray_Enumerable_OddBeforeEven(IReadOnlyList<int> data, ImmutableArray<int> expected)
-    {
-        var enumerable = (IEnumerable<int>)data;
-        var sorted = enumerable.OrderAsArray(OddBeforeEven);
-        Assert.Equal<int>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderDescendingTestData))]
-    public void OrderDescendingAsArray_Enumerable(IReadOnlyList<int> data, ImmutableArray<int> expected)
-    {
-        var enumerable = (IEnumerable<int>)data;
-        var sorted = enumerable.OrderDescendingAsArray();
-        Assert.Equal<int>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderDescendingTestData_OddBeforeEven))]
-    public void OrderDescendingAsArray_Enumerable_OddBeforeEven(IReadOnlyList<int> data, ImmutableArray<int> expected)
-    {
-        var enumerable = (IEnumerable<int>)data;
-        var sorted = enumerable.OrderDescendingAsArray(OddBeforeEven);
-        Assert.Equal<int>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderByTestData))]
-    public void OrderByAsArray_Enumerable(IReadOnlyList<ValueHolder> data, ImmutableArray<ValueHolder> expected)
-    {
-        var enumerable = (IEnumerable<ValueHolder>)data;
-        var sorted = enumerable.OrderByAsArray(static x => x.Value);
-        Assert.Equal<ValueHolder>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderByTestData_OddBeforeEven))]
-    public void OrderByAsArray_Enumerable_OddBeforeEven(IReadOnlyList<ValueHolder> data, ImmutableArray<ValueHolder> expected)
-    {
-        var enumerable = (IEnumerable<ValueHolder>)data;
-        var sorted = enumerable.OrderByAsArray(static x => x.Value, OddBeforeEven);
-        Assert.Equal<ValueHolder>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderByDescendingTestData))]
-    public void OrderByDescendingAsArray_Enumerable(IReadOnlyList<ValueHolder> data, ImmutableArray<ValueHolder> expected)
-    {
-        var enumerable = (IEnumerable<ValueHolder>)data;
-        var sorted = enumerable.OrderByDescendingAsArray(static x => x.Value);
-        Assert.Equal<ValueHolder>(expected, sorted);
-    }
-
-    [Theory]
-    [MemberData(nameof(OrderByDescendingTestData_OddBeforeEven))]
-    public void OrderByDescendingAsArray_Enumerable_OddBeforeEven(IReadOnlyList<ValueHolder> data, ImmutableArray<ValueHolder> expected)
-    {
-        var enumerable = (IEnumerable<ValueHolder>)data;
-        var sorted = enumerable.OrderByDescendingAsArray(static x => x.Value, OddBeforeEven);
-        Assert.Equal<ValueHolder>(expected, sorted);
-    }
-
-#if NET // Enumerable.Order(...) and Enumerable.OrderDescending(...) were introduced in .NET 7
-
-    [Fact]
-    public void OrderAsArray_IsStable()
-    {
-        OrderAndAssertStableSort(
-            linqFunction: data => data.Order(),
-            testFunction: data => data.OrderAsArray());
-    }
-
-    [Fact]
-    public void OrderAsArray_Comparer_IsStable()
-    {
-        OrderAndAssertStableSort(
-            linqFunction: data => data.Order(StringHolder.Comparer.Ordinal),
-            testFunction: data => data.OrderAsArray(StringHolder.Comparer.Ordinal));
-
-        OrderAndAssertStableSort(
-            linqFunction: data => data.Order(StringHolder.Comparer.OrdinalIgnoreCase),
-            testFunction: data => data.OrderAsArray(StringHolder.Comparer.OrdinalIgnoreCase));
-    }
-
-    [Fact]
-    public void OrderDescendingAsArray_IsStable()
-    {
-        OrderAndAssertStableSort(
-            linqFunction: data => data.OrderDescending(),
-            testFunction: data => data.OrderDescendingAsArray());
-    }
-
-    [Fact]
-    public void OrderDescendingAsArray_Comparer_IsStable()
-    {
-        OrderAndAssertStableSort(
-            linqFunction: data => data.OrderDescending(StringHolder.Comparer.Ordinal),
-            testFunction: data => data.OrderDescendingAsArray(StringHolder.Comparer.Ordinal));
-
-        OrderAndAssertStableSort(
-            linqFunction: data => data.OrderDescending(StringHolder.Comparer.OrdinalIgnoreCase),
-            testFunction: data => data.OrderDescendingAsArray(StringHolder.Comparer.OrdinalIgnoreCase));
-    }
-
-#endif
-
-    [Fact]
-    public void OrderByAsArray_IsStable()
-    {
-        OrderAndAssertStableSort(
-            linqFunction: data => data.OrderBy(static x => x?.Text),
-            testFunction: data => data.OrderByAsArray(static x => x?.Text));
-    }
-
-    [Fact]
-    public void OrderByAsArray_Comparer_IsStable()
-    {
-        OrderAndAssertStableSort(
-            linqFunction: data => data.OrderBy(static x => x?.Text, StringComparer.OrdinalIgnoreCase),
-            testFunction: data => data.OrderByAsArray(static x => x?.Text, StringComparer.OrdinalIgnoreCase));
-
-        OrderAndAssertStableSort(
-            linqFunction: data => data.OrderBy(static x => x?.Text, StringComparer.OrdinalIgnoreCase),
-            testFunction: data => data.OrderByAsArray(static x => x?.Text, StringComparer.OrdinalIgnoreCase));
-    }
-
-    [Fact]
-    public void OrderByDescendingAsArray_IsStable()
-    {
-        OrderAndAssertStableSort(
-            linqFunction: data => data.OrderByDescending(static x => x?.Text),
-            testFunction: data => data.OrderByDescendingAsArray(static x => x?.Text));
-    }
-
-    [Fact]
-    public void OrderByDescendingAsArray_Comparer_IsStable()
-    {
-        OrderAndAssertStableSort(
-            linqFunction: data => data.OrderByDescending(static x => x?.Text, StringComparer.OrdinalIgnoreCase),
-            testFunction: data => data.OrderByDescendingAsArray(static x => x?.Text, StringComparer.OrdinalIgnoreCase));
-
-        OrderAndAssertStableSort(
-            linqFunction: data => data.OrderByDescending(static x => x?.Text, StringComparer.OrdinalIgnoreCase),
-            testFunction: data => data.OrderByDescendingAsArray(static x => x?.Text, StringComparer.OrdinalIgnoreCase));
-    }
-
-    private static void OrderAndAssertStableSort(
-        Func<IReadOnlyList<StringHolder?>, IEnumerable<StringHolder?>> linqFunction,
-        Func<IReadOnlyList<StringHolder?>, ImmutableArray<StringHolder?>> testFunction)
-    {
-        IReadOnlyList<StringHolder?> data = [
-            "All", "Your", "Base", "Are", "belong", null, "To", "Us",
-            "all", "your", null, "Base", "are", "belong", "to", "us"];
-
-        var expected = linqFunction(data);
-        var actual = testFunction(data);
-
-        Assert.Equal<StringHolder?>(expected, actual, ReferenceEquals);
-    }
 }
