+++ src/Shared/Microsoft.AspNetCore.Razor.Utilities.Shared.Test/TestData/EnumerableOrderingTestBase.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+using System.Collections.Generic;
+
+namespace Microsoft.AspNetCore.Razor.Utilities.Shared.Test.TestData;
+
+public abstract class EnumerableOrderingTestBase : OrderingTestBase<IEnumerable<int>, IEnumerable<ValueHolder<int>>, OrderingCaseConverters.Enumerable>
+{
+}
